Add files via upload
This report will analyze the Glasgow Norms dataset, the result of an experiment in psycholinguistics that analyzes polysemic words, i.e. words that, despite being homographical, are characterized by different meanings, describing them by means of nine different attributes. In this project we will try to analyze and understand if polysemy can be recognized in an automatic way by means of clustering, classification and pattern mining algorithms and what are the characteristics that make these words recognizable.